requirements.txt
altgraph==0.10.2
bdist-mpkg==0.5.0
certifi==2017.7.27.1
chardet==3.0.4
modulegraph==0.10.4
oauthlib==2.0.2
py2app==0.7.3
pyparsing==2.0.1
python-dateutil==1.5
python-twitter==3.3
pytz==2013.7
requests==2.18.4
requests-oauthlib==0.8.0
six==1.4.1
tweetpy==0.1
urllib3==1.22
virtualenv==15.1.0
bottle==0.12.13
